properly define mpl-1.1
"foo or bar" in dep5 has the actual meaning of this license or that license
where each of them is a standalone thing that needs to be defined as
a license block. to represen this properly we now have an mpl-1.1 block
and a gpl-2+ block.

should make lintian happier